[FIX] * : Allow public access to custom product placeholders
* : loyalty, website_sale_loyalty, website_event_sale

Placeholders were introduced in c52453f for meta products such as
discounts, gift cards, event tickets, and appointments.

Previously, access rights prevented public users from viewing these
placeholders. Even when no image was linked to a record, access was
denied due to read restrictions, defaulting to the generic placeholder
instead of the custom one provided by `_get_product_placeholder_filename`.

This commit overrides `_can_return_content` in relevant modules to allow
proper access for public users, ensuring the intended placeholder is
displayed.
