+++ addons/website_event_sale/models/product.py
@@ -9,6 +9,24 @@ class ProductProduct(models.Model):
 
     event_ticket_ids = fields.One2many('event.event.ticket', 'product_id', string='Event Tickets')
 
+    def _can_return_content(self, field_name=None, access_token=None):
+        """ Override of `orm` to give public users access to the unpublished product image.
+
+        Give access to the public users to the unpublished product images if they are linked to an
+        event ticket.
+
+        :param field_name: The name of the field to check.
+        :param access_token: The access token.
+        :return: Whether to allow the access to the image.
+        :rtype: bool
+        """
+        if (
+            field_name in ["image_%s" % size for size in [1920, 1024, 512, 256, 128]]
+            and self.sudo().event_ticket_ids
+        ):
+            return True
+        return super()._can_return_content(field_name, access_token)
+
     def _get_product_placeholder_filename(self):
         if self.event_ticket_ids:
             return 'website_event_sale/static/img/event_ticket_placeholder_thumbnail.png'
