+++ addons/website_sale_loyalty/models/__init__.py
@@ -4,5 +4,6 @@
 from . import loyalty_card
 from . import loyalty_program
 from . import loyalty_rule
+from . import product_product
 from . import sale_order_line
 from . import sale_order
